refactor: 컨트롤러 및 테스트 분리
