+++ src/main/java/com/withme/api/controller/UserController.java
@@ -2,8 +2,6 @@ package com.withme.api.controller;
 
 import com.withme.api.controller.dto.ExceptionResponseDto;
 import com.withme.api.controller.dto.JoinRequestDto;
-import com.withme.api.controller.dto.MyPageResponseDto;
-import com.withme.api.controller.dto.UserUpdateRequestDto;
 import com.withme.api.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -14,7 +12,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
 
@@ -52,65 +53,6 @@ public class UserController {
         userService.createUser(dto);
 
         return new ResponseEntity<>(HttpStatus.CREATED);
-    }
-
-    @Operation(
-            summary = "닉네임 변경"
-            , description = "유저의 닉네임을 수정한다."
-    )
-    @ApiResponses(value = {
-            @ApiResponse(
-                responseCode = "200"
-                , description = "닉네임 변경 성공"
-            )
-            , @ApiResponse(
-                responseCode = "422"
-                , description = "파라미터 유효성 부적합"
-                , content = @Content(schema = @Schema(implementation = ExceptionResponseDto.class))
-            )
-            , @ApiResponse(
-                responseCode = "400"
-                , description = "이메일 혹은 닉네임 중복"
-                , content = @Content(schema = @Schema(implementation = ExceptionResponseDto.class))
-            )
-            , @ApiResponse(
-                responseCode = "400"
-                , description = "id에 일치하는 유저 없음"
-                , content = @Content(schema = @Schema(implementation = ExceptionResponseDto.class))
-            )
-    })
-    @PutMapping("/user/nickname/{id}")
-    public ResponseEntity<Object> changeUserNickname(@PathVariable Long id, @Valid @RequestBody UserUpdateRequestDto dto) {
-        log.debug("changeUserNickname{} invoked", dto);
-        userService.changeUserNickname(id, dto);
-        return ResponseEntity.ok().build();
-    }
-
-    @Operation(
-        summary = "마이페이지 유저 및 팀 정보 조회"
-        , description = "마이페이지에서 본인의 닉네임과 속해있는 팀 정보를 조회한다."
-    )
-    @ApiResponses(value = {
-        @ApiResponse(
-            responseCode = "200"
-            , description = "조회 성공"
-            , content = @Content(schema = @Schema(implementation = MyPageResponseDto.class))
-
-        )
-        , @ApiResponse(
-            responseCode = "400"
-            , description = "id에 일치하는 유저 없음"
-            , content = @Content(schema = @Schema(implementation = ExceptionResponseDto.class))
-        )
-    })
-    @GetMapping("/user/mypage/{id}")
-    public ResponseEntity<Object> getUserAndTeamInfo(@PathVariable Long id) {
-        log.debug("getUserAndTeamInfo{} invoked", id);
-        MyPageResponseDto myPageResponseDto = userService.getUserAndTeamInfo(id);
-
-        log.debug("myPageResponseDto : " + myPageResponseDto);
-
-        return ResponseEntity.ok().body(myPageResponseDto);
     }
 
 }
